add go mod to provioner server build
