+++ ansible/provisioner_server/Dockerfile
@@ -1,7 +1,7 @@
 FROM golang as builder
 WORKDIR /app
 COPY . .
-RUN go build -o server -ldflags "-w -s" server.go
+RUN go build -o server -ldflags "-w -s" .
 FROM ubuntu:20.04
 EXPOSE 5555
 WORKDIR /home/ubuntu
