+++ ansible/provisioner_server/go.mod
@@ -1,0 +1,3 @@
+module provisioner/server
+
+go 1.16
